pyproject.toml
[project]
name = "text-sql-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kagglehub>=0.4.0",
    "langgraph",
    "supabase",
    "pandas",
    "streamlit",
    "python-dotenv",
    "plotly",
    "psycopg2-binary",
    "google-genai>=1.60.0",
]
